group tests for old_style in one file
